requirements.txt
#requirements.txt

djangorestframework
psycopg2-binary
python-decouple
Pillow
django-filter
djangorestframework-jwt
djangorestframework-simplejwt
djangorestframework-simplejwt
coverage

